structure search results as a tree
