+++ sites/kit.svelte.dev/src/lib/search/types.d.ts
@@ -4,3 +4,10 @@ export interface Block {
 	content: string;
 	rank: number;
 }
+
+export interface Tree {
+	breadcrumbs: string[];
+	href: string;
+	node: Block;
+	children: Tree[];
+}
